Also branch-select VITE_CONVEX_URL for Cloudflare preview deploys
Same problem as the Clerk key, one layer deeper: VITE_CONVEX_URL was
also a single flat Cloudflare build variable, so preview deploys pointed
at the production Convex deployment even after the Clerk key fix landed.
That deployment's CLERK_JWT_ISSUER_DOMAIN only trusts the production
Clerk instance, so it rejected the (correctly issued) dev-instance
tokens a preview build now presents - the Convex WebSocket auth
handshake failed in a reconnect loop instead ("Loading..." forever,
console showing repeated "WebSocket reconnected ... closed with code
1005"). Confirmed via `npx convex env list` / `--prod` that the dev
deployment (keen-marmot-770) trusts the dev Clerk issuer
(adjusted-bug-38.clerk.accounts.dev) while production
(qualified-terrier-968) trusts the production one (clerk.p2b.butaud.com).

select-clerk-key.js is renamed to select-preview-env.js and its
override logic is generalized to handle both VITE_CLERK_PUBLISHABLE_KEY
and VITE_CONVEX_URL identically, each independently activated by its
own _PROD/_PREVIEW pair. Verified locally for both branch cases with
fake values, confirming the correct URL/key ends up in the compiled
bundle.
